+++ bazelrunner/src/main/java/org/jetbrains/bsp/bazel/bazelrunner/BazelRunner.java
@@ -7,8 +7,7 @@ import java.io.IOException;
 import java.util.List;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.jetbrains.bsp.bazel.bazelrunner.data.BazelData;
-import org.jetbrains.bsp.bazel.server.loggers.BuildClientLogger;
+import org.jetbrains.bsp.bazel.logger.BuildClientLogger;
 
 public class BazelRunner {
 
@@ -20,21 +19,24 @@ public class BazelRunner {
   private final String bazel;
 
   private Option<Integer> besBackendPort = Option.none();
-  private Option<BuildClientLogger> buildClientLogger = Option.none();
+  private final BuildClientLogger buildClientLogger;
   private final BazelData bazelData;
 
   // This is runner without workspace path. It is used to determine workspace
   // path and create a fully functional runner.
-  public static BazelRunner inCwd(String bazelBinaryPath) {
-    return new BazelRunner(bazelBinaryPath, null);
+  public static BazelRunner inCwd(String bazelBinaryPath, BuildClientLogger buildClientLogger) {
+    return new BazelRunner(bazelBinaryPath, buildClientLogger, null);
   }
 
-  public static BazelRunner of(String bazelBinaryPath, BazelData bazelData) {
-    return new BazelRunner(bazelBinaryPath, bazelData);
+  public static BazelRunner of(
+      String bazelBinaryPath, BuildClientLogger buildClientLogger, BazelData bazelData) {
+    return new BazelRunner(bazelBinaryPath, buildClientLogger, bazelData);
   }
 
-  private BazelRunner(String bazelBinaryPath, BazelData bazelData) {
+  private BazelRunner(
+      String bazelBinaryPath, BuildClientLogger buildClientLogger, BazelData bazelData) {
     this.bazel = bazelBinaryPath;
+    this.buildClientLogger = buildClientLogger;
     this.bazelData = bazelData;
   }
 
@@ -43,13 +45,12 @@ public class BazelRunner {
   }
 
   BazelProcess runBazelCommandBes(String command, List<String> flags, List<String> arguments) {
-    var newFlags = getBesFlags(flags);
-
+    List<String> newFlags = getBesFlags(flags);
     return runBazelCommand(command, newFlags, arguments);
   }
 
   private List<String> getBesFlags(List<String> flags) {
-    var newFlags = Lists.newArrayList(getBesBackendAddress(), PUBLISH_ALL_ACTIONS);
+    List<String> newFlags = Lists.newArrayList(getBesBackendAddress(), PUBLISH_ALL_ACTIONS);
     newFlags.addAll(flags);
 
     return newFlags;
@@ -57,15 +58,11 @@ public class BazelRunner {
 
   private String getBesBackendAddress() {
     var port = besBackendPort.getOrElseThrow(() -> new IllegalStateException("BES port not set"));
-
     return BES_BACKEND + port;
   }
 
   BazelProcess runBazelCommand(String command, List<String> flags, List<String> arguments) {
-
     try {
-      LOGGER.info(
-          "Waiting for bazel - command: {}, flags: {}, arguments: {}", command, flags, arguments);
       return runBazelProcess(command, flags, arguments);
     } catch (IOException e) {
       throw new RuntimeException(e);
@@ -75,7 +72,10 @@ public class BazelRunner {
   private synchronized BazelProcess runBazelProcess(
       String command, List<String> flags, List<String> arguments) throws IOException {
     var processArgs = getProcessArgs(command, flags, arguments);
-    LOGGER.info("Running: {}", String.join(" ", processArgs));
+
+    var message = "Invoking: " + String.join(" ", processArgs);
+    LOGGER.info(message);
+    buildClientLogger.logMessage(message);
 
     var processBuilder = new ProcessBuilder(processArgs);
     if (bazelData != null) {
@@ -90,15 +90,10 @@ public class BazelRunner {
     var processArgs = Lists.newArrayList(bazel, command);
     processArgs.addAll(flags);
     processArgs.addAll(arguments);
-
     return processArgs;
   }
 
   public void setBesBackendPort(int port) {
     besBackendPort = Option.of(port);
-  }
-
-  public void setLogger(BuildClientLogger buildClientLogger) {
-    this.buildClientLogger = Option.of(buildClientLogger);
   }
 }