+++ logger/src/main/java/org/jetbrains/bsp/bazel/logger/BuildClientLogger.java
@@ -1,4 +1,4 @@
-package org.jetbrains.bsp.bazel.server.loggers;
+package org.jetbrains.bsp.bazel.logger;
 
 import ch.epfl.scala.bsp4j.BuildClient;
 import ch.epfl.scala.bsp4j.LogMessageParams;
@@ -6,28 +6,26 @@ import ch.epfl.scala.bsp4j.MessageType;
 
 public class BuildClientLogger {
 
-  private final BuildClient buildClient;
-
-  public BuildClientLogger(BuildClient buildClient) {
-    this.buildClient = buildClient;
-  }
+  private BuildClient buildClient;
 
   public void logError(String errorMessage) {
-    logIfNotBlank(MessageType.ERROR, errorMessage);
+    log(MessageType.ERROR, errorMessage);
   }
 
   public void logMessage(String message) {
-    logIfNotBlank(MessageType.LOG, message);
-  }
-
-  private void logIfNotBlank(MessageType messageType, String message) {
-    if (!message.trim().isEmpty()) {
-      log(messageType, message);
-    }
+    log(MessageType.LOG, message);
   }
 
   private void log(MessageType messageType, String message) {
-    LogMessageParams params = new LogMessageParams(messageType, message.trim());
-    buildClient.onBuildLogMessage(params);
+    if (buildClient == null) return;
+
+    if (!message.trim().isEmpty()) {
+      var params = new LogMessageParams(messageType, message);
+      buildClient.onBuildLogMessage(params);
+    }
+  }
+
+  public void setBuildClient(BuildClient buildClient) {
+    this.buildClient = buildClient;
   }
 }