+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+
 # Changelog
 
 <!-- Keep a Changelog guide -> https://keepachangelog.com -->
@@ -6,6 +7,9 @@
 
 ### Features 🎉
 
+- Improved bazel runner. The BSP Client now receives invoked bazel command,
+  its output, duration and exit code.
+  | [#198](https://github.com/JetBrains/bazel-bsp/pull/198)
 - Scala diagnostics work without a fork (technically it worked already for a while).
   | [#170](https://github.com/JetBrains/bazel-bsp/pull/170)